fix(fnp-ufunc): widen test-data multiplies in par_select tests to avoid debug i32 overflow
par_select_median_matches_serial_across_distributions and
par_select_percentile_matches_serial_across_fractions_and_methods failed
deterministically with "attempt to multiply with overflow" in debug builds. Three
data-generator closures multiplied an untyped loop index (which defaults to i32) by
a large constant: `i * 40503` over ranges of 100k/200k (up to 8.1e9) and the LCG
`i * 1103515245 + 12345` over 262k (~2.9e14). These overflow i32 and panic under
`cargo test` (debug), while release wraps silently — so the suite was red in debug
and only green in release. Cast the index to u64 before the multiply (`i as u64 *
...`) so the arithmetic is well-defined in both modes. The tests are parallel-vs-
serial self-consistency checks on identical data, so the concrete values are
immaterial; both tests now pass in debug.

deadlock-audit-r9sml

+++ crates/fnp-ufunc/src/lib.rs
@@ -48191,7 +48191,7 @@ print(json.dumps(payload))
                 .map(|i| ((i * 2654435761u64 as usize) % 100003) as f64 * 1e-2 - 500.0)
                 .collect(),
             (0..200000)
-                .map(|i| ((i * 40503) % 977) as f64 - 488.0)
+                .map(|i| ((i as u64 * 40503) % 977) as f64 - 488.0)
                 .collect(),
             (0..150001).map(|i| i as f64).collect(),
             (0..150000).map(|i| (150000 - i) as f64).collect(),
@@ -48245,7 +48245,7 @@ print(json.dumps(payload))
                 .map(|i| ((i * 2654435761u64 as usize) % 97) as f64)
                 .collect(),
             (0..100000)
-                .map(|i| ((i * 40503) % 131) as f64 - 60.0)
+                .map(|i| ((i as u64 * 40503) % 131) as f64 - 60.0)
                 .collect(),
             (0..200003).map(|i| i as f64).collect(),
             (0..200000).map(|i| (200000 - i) as f64).collect(),
@@ -48267,7 +48267,7 @@ print(json.dumps(payload))
                 .map(|i| ((i * 2246822519u32 as usize) % 100003) as f64 * 1e-3 - 50.0)
                 .collect(),
             (0..262145)
-                .map(|i| (((i * 1103515245 + 12345) % 1000) as f64) - 500.0)
+                .map(|i| (((i as u64 * 1103515245 + 12345) % 1000) as f64) - 500.0)
                 .collect(),
         ];
         for data in &cases {
